+++ src/favs/favs.controller.ts
@@ -1,6 +1,5 @@
 import { Controller, Get, Post, Body, Param, Delete } from '@nestjs/common';
 import { FavsService } from './favs.service';
-import { CreateFavDto } from './dto/create-fav.dto';
 
 @Controller('favs')
 export class FavsController {
