+++ src/user/user.controller.ts
@@ -10,7 +10,7 @@ import {
   Put,
 } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
-import { UpdatePasswordDto } from './dto/update-user-password.dto';
+import { UpdatePasswordDto } from './dto/update-user.dto';
 import { UserService } from './user.service';
 
 @Controller('user')
@@ -18,19 +18,16 @@ export class UserController {
   constructor(private readonly UserService: UserService) {}
 
   @Get()
-  @HttpCode(HttpStatus.OK)
   getUsers() {
     return this.UserService.getUsers();
   }
 
   @Get(':id')
-  @HttpCode(HttpStatus.OK)
   getUserById(@Param('id') id: string) {
     return this.UserService.getUserById(id);
   }
 
   @Post()
-  @HttpCode(HttpStatus.CREATED)
   CreateUserDto(@Body() CreateUserDto: CreateUserDto) {
     return this.UserService.CreateUserDto(CreateUserDto);
   }
